Creating the JUnit DAO testing
The way to test a DAO is to test all of its method against your
database. This helps you tackle each individual method and check if it
has been implemented correctly.

Note: Remember to clear the test db before performing any tests. Not
necessary but is helpful for a basic db.

Important methods: assertEquals, assertTrue